Remove stories and moves component to devetools panel
