--- packages/extension/src/view/devtools/components/tabContent/cookies/components/cookieList.tsx
@@ -1,58 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- * External dependencies.
- */
-import React from 'react';
-
-/**
- * Internal dependencies.
- */
-import { type CookieData } from '../../../../../../localStore';
-import ListItem from './listItem';
-
-interface CoookieListProps {
-  cookies: {
-    [key: string]: CookieData;
-  };
-  tabUrl: string | null;
-  selectedKey: string | null;
-  onClickItem: (key: string) => void;
-}
-
-const CookieList = ({
-  cookies,
-  tabUrl,
-  selectedKey,
-  onClickItem,
-}: CoookieListProps) => (
-  <ul className="w-full h-full">
-    {cookies &&
-      tabUrl &&
-      Object.entries(cookies).map(([key, value]) => (
-        <li key={key}>
-          <ListItem
-            cookie={value}
-            tabUrl={tabUrl}
-            isSelected={selectedKey === key}
-            onClick={() => onClickItem(key)}
-          />
-        </li>
-      ))}
-  </ul>
-);
-
-export default CookieList;
