+++ packages/extension/src/view/devtools/components/table/tableHeader/headerResizer.tsx
